Fixed merge issues from other branch
This contains the PR from isometric-background, but I had to make this branch to manage the merge.

I also added an Area2D to replace the programatic bounds for the background baseballs, but I wasn't sure how to tie the signal emitter for area exit to the preload bgBaseball scene, so I will do that later.